refactor(classic): move contents of download/classic to resource/classic
This prepares a refactoring PR with the goal of making the download code independent of the concrete feature/API
