+++ pkg/resource/classic/doc.go
