+++ pkg/resource/classic/download_sanitize.go
